% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_where_clause.R
\name{build_where_clause}
\alias{build_where_clause}
\title{Build Where Clause}
\usage{
build_where_clause(
  tab_name,
  col_name,
  col_value,
  con = getOption("podr_connection")
)
}
\arguments{
\item{tab_name}{table name}

\item{col_name}{column name}

\item{col_value}{column value}

\item{con}{The connection to PODR. Use connect_podr to establish a connection, or specify a variable
containing the proper PostgreSQL connection into PODR yourself}
}
\value{
column definition in a data.frame
}
\description{
This function builds a where clause for a SQL statement
}
\examples{
\dontrun{
cc <- conn_podr()
build_where_clause('css_2020_ae',con=cc)
}

}
\author{
Hanming Tu
}
