% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanpod.R
\name{cleanpod}
\alias{cleanpod}
\title{Given a podcast data frame created with readpod, this function will return
a data frame that contains four variables:
#' \enumerate{
  \item podtitle (podcast title)
  \item title (episode title)
  \item date (publish date)
  \item showlength (show length in seconds)
}}
\usage{
cleanpod(dfreadpod)
}
\arguments{
\item{dfreadpod}{a data frame created with readpod}
}
\value{
cleaned episode data contained in \code{dfreadpod}
}
\description{
Given a podcast data frame created with readpod, this function will return
a data frame that contains four variables:
#' \enumerate{
  \item podtitle (podcast title)
  \item title (episode title)
  \item date (publish date)
  \item showlength (show length in seconds)
}
}
\note{
Every podcast feed is not created equally and some feeds may
be missing data. Be on the lookout for the following issues:
\itemize{
  \item episode as announcement == duration of 00 or NA
  \item duration not entered in <itunes:duration> results in empty character
  vector
  \item episode durations in <itunes:duration> may be in HH:MM:SS and
  some may be in H:MM:SS, and still others in just MM:SS. This function should
  be able to handle each.
  \item duplication of episodes in the feed
}
}
\examples{
{
cleanpod(jjgo)
}
}
