% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table_defs.R
\name{get_table_defs}
\alias{get_table_defs}
\title{Get Table Definition from PODR}
\usage{
get_table_defs(tab_name, con = getOption("podr_connection"))
}
\arguments{
\item{tab_name}{table name}

\item{con}{The connection to PODR. Use connect_podr to establish a connection, or specify a variable
containing the proper PostgreSQL connection into PODR yourself}
}
\value{
column definition in a data.frame
}
\description{
This function will get column definition for a table in"PODR".
}
\examples{
\dontrun{
cc <- conn_podr()
get_table_defs('css_2020_ae',con=cc)
}

}
\author{
Hanming Tu
}
