% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_app.R
\name{start_app}
\alias{start_app}
\title{Start R Shiny app}
\usage{
start_app(
  app_name = "01_podr",
  n = 1,
  pkg = "podr",
  pt = NULL,
  lb = getOption("shiny.launch.browser", interactive()),
  ht = getOption("shiny.host", "127.0.0.1"),
  dm = "normal",
  msg_lvl = NULL,
  loc = "local"
)
}
\arguments{
\item{app_name}{app or script name}

\item{n}{app number}

\item{pkg}{package name}

\item{pt}{Port number}

\item{lb}{define the browser- shiny.launch.browser}

\item{ht}{define the host or ip address}

\item{dm}{display modes are auto, normal or showcase}

\item{msg_lvl}{message level}

\item{loc}{location of the scirpt: local|github; default to 'local'}
}
\description{
start R Shiny apps included in this package.
}
\examples{
\dontrun{
  library(genTS)
  start_appe()  # default to "02_display"
  start_app(1)  # start "01_html"
}
}
\author{
Hanming Tu
}
