% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_webapp.R
\name{start_webapp}
\alias{start_webapp}
\title{Start Web Application}
\usage{
start_webapp(
  n = 1,
  pkg = "podr",
  pt = NULL,
  lb = getOption("shiny.launch.browser", interactive()),
  ht = getOption("shiny.host", "127.0.0.1"),
  dm = "normal",
  msg_lvl = NULL
)
}
\arguments{
\item{n}{Example number}

\item{pkg}{package name}

\item{pt}{Port number}

\item{lb}{define the browser- shiny.launch.browser}

\item{ht}{define the host or ip address}

\item{dm}{display modes are auto, normal or showcase}

\item{msg_lvl}{message display level}
}
\description{
start web appllication framework.
}
\examples{
  # library(bldsql)
  # comment out the interactive sessions
  # start_webapp()  # default to "01_webapp"
}
\author{
Hanming Tu
}
